@@ -20,11 +20,12 @@ exclude = [
   "/pkg/brew",
   "/benchsuite/",
   "/scripts/",
+  "/crates/fuzz",
 ]
 build = "build.rs"
 autotests = false
-edition = "2021"
-rust-version = "1.70"
+edition = "2024"
+rust-version = "1.85"
 
 [[bin]]
 bench = false
@@ -64,8 +65,8 @@ version = "1.3.0"
 default-features = false
 features = ["std"]
 
-[target.'cfg(all(target_env = "musl", target_pointer_width = "64"))'.dependencies.jemallocator]
-version = "0.5.0"
+[target.'cfg(all(target_env = "musl", target_pointer_width = "64"))'.dependencies.tikv-jemallocator]
+version = "0.6.0"
 
 [dev-dependencies]
 serde = "1.0.77"
