@@ -13,7 +13,7 @@ repository = "https://github.com/BurntSushi/ripgrep/tree/master/crates/globset"
 readme = "README.md"
 keywords = ["regex", "glob", "multiple", "set", "pattern"]
 license = "Unlicense OR MIT"
-edition = "2021"
+edition = "2024"
 
 [lib]
 name = "globset"
@@ -21,6 +21,7 @@ bench = false
 
 [dependencies]
 aho-corasick = "1.1.1"
+arbitrary = { version = "1.3.2", optional = true, features = ["derive"] }
 bstr = { version = "1.6.2", default-features = false, features = ["std"] }
 log = { version = "0.4.20", optional = true }
 serde = { version = "1.0.188", optional = true }
@@ -41,6 +42,7 @@ serde_json = "1.0.107"
 
 [features]
 default = ["log"]
+arbitrary = ["dep:arbitrary"]
 # DEPRECATED. It is a no-op. SIMD is done automatically through runtime
 # dispatch.
 simd-accel = []
