@@ -12,7 +12,7 @@ repository = "https://github.com/BurntSushi/ripgrep/tree/master/crates/ignore"
 readme = "README.md"
 keywords = ["glob", "ignore", "gitignore", "pattern", "file"]
 license = "Unlicense OR MIT"
-edition = "2021"
+edition = "2024"
 
 [lib]
 name = "ignore"
@@ -41,7 +41,7 @@ version = "0.1.2"
 
 [dev-dependencies]
 bstr = { version = "1.6.2", default-features = false, features = ["std"] }
-crossbeam-channel = "0.5.8"
+crossbeam-channel = "0.5.15"
 
 [features]
 # DEPRECATED. It is a no-op. SIMD is done automatically through runtime
