@@ -4,8 +4,8 @@ use std::{
     fs::{self, FileType, Metadata},
     io,
     path::{Path, PathBuf},
-    sync::atomic::{AtomicBool, AtomicUsize, Ordering as AtomicOrdering},
     sync::Arc,
+    sync::atomic::{AtomicBool, AtomicUsize, Ordering as AtomicOrdering},
 };
 
 use {
@@ -15,11 +15,11 @@ use {
 };
 
 use crate::{
+    Error, PartialErrorBuilder,
     dir::{Ignore, IgnoreBuilder},
     gitignore::GitignoreBuilder,
     overrides::Override,
     types::Types,
-    Error, PartialErrorBuilder,
 };
 
 /// A directory entry with a possible error attached.
@@ -809,6 +809,10 @@ impl WalkBuilder {
     ///
     /// When disabled, git-related ignore rules are applied even when searching
     /// outside a git repository.
+    ///
+    /// In particular, if this is `false` then `.gitignore` files will be read
+    /// from parent directories above the git root directory containing `.git`,
+    /// which is different from the git behavior.
     pub fn require_git(&mut self, yes: bool) -> &mut WalkBuilder {
         self.ig_builder.require_git(yes);
         self
@@ -905,6 +909,10 @@ impl WalkBuilder {
     ///
     /// Note that the errors for reading entries that may not satisfy the
     /// predicate will still be yielded.
+    ///
+    /// Note also that only one filter predicate can be applied to a
+    /// `WalkBuilder`. Calling this subsequent times overrides previous filter
+    /// predicates.
     pub fn filter_entry<P>(&mut self, filter: P) -> &mut WalkBuilder
     where
         P: Fn(&DirEntry) -> bool + Send + Sync + 'static,
@@ -1316,7 +1324,7 @@ impl WalkParallel {
 
     fn threads(&self) -> usize {
         if self.threads == 0 {
-            2
+            std::thread::available_parallelism().map_or(1, |n| n.get()).min(12)
         } else {
             self.threads
         }
@@ -1431,8 +1439,11 @@ impl Stack {
                 stealers: stealers.clone(),
             })
             .collect();
-        // Distribute the initial messages.
+        // Distribute the initial messages, reverse the order to cancel out
+        // the other reversal caused by the inherent LIFO processing of the
+        // per-thread stacks which are filled here.
         init.into_iter()
+            .rev()
             .zip(stacks.iter().cycle())
             .for_each(|(m, s)| s.push(m));
         stacks
@@ -1898,7 +1909,7 @@ fn device_num<P: AsRef<Path>>(path: P) -> io::Result<u64> {
 
 #[cfg(windows)]
 fn device_num<P: AsRef<Path>>(path: P) -> io::Result<u64> {
-    use winapi_util::{file, Handle};
+    use winapi_util::{Handle, file};
 
     let h = Handle::from_path_any(path)?;
     file::information(h).map(|info| info.volume_serial_number())
@@ -1944,11 +1955,7 @@ mod tests {
     }
 
     fn normal_path(unix: &str) -> String {
-        if cfg!(windows) {
-            unix.replace("\\", "/")
-        } else {
-            unix.to_string()
-        }
+        if cfg!(windows) { unix.replace("\\", "/") } else { unix.to_string() }
     }
 
     fn walk_collect(prefix: &Path, builder: &WalkBuilder) -> Vec<String> {
